avutil/tx: make tx table configurable at build time
Commit 8f48a62 extends tx to 2M, which takes 4M of bss section for
tx_float:

size libavutil/tx_float.o
   text    data     bss     dec     hex filename
  35356    9160 4194796 4239312  40afd0 libavutil/tx_float.o

This isn't a issue on devices with normal memory sizes and OS
supporting virtual memory. But it's a real issue for embedded devices
with realtime OS, which may not support virtual memory, e.g., Nuttx.
This 4M of bss section will occupy physical memory, which is a scarce
resource on embedded devices.